feat(compaction,2.3): add some unit tests
