+++ tskv/src/compaction/flush.rs
@@ -315,26 +315,6 @@ impl WriterWrapper {
             }
         };
 
-        // // Split values for level-0 and levle-1.
-        // let splited_values = match values.binary_search_by(|v| v.0.cmp(&self.max_level_ts)) {
-        //     Ok(i) => {
-        //         if i == values.len() - 1 {
-        //             [values, &[]]
-        //         } else {
-        //             [&values[..=i], &values[(i + 1)..]]
-        //         }
-        //     }
-        //     Err(i) => {
-        //         if i == 0 {
-        //             [&[], values]
-        //         } else if i < values.len() {
-        //             [&values[..i], &values[i..]]
-        //         } else {
-        //             [values, &[]]
-        //         }
-        //     }
-        // };
-
         // flush all data to delta level
         // Fill buffer and write to disk if buffer is full.
         for (ts, val) in values {
